backend/pyproject.toml
[project]
name = "aurelia-backend"
version = "0.1.0"
description = "Aurelia local-first backend service"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.14.0",
  "fastapi>=0.115.0",
  "pgembed>=0.1.9",
  "psycopg[binary]>=3.2.0",
  "python-dotenv>=1.0.1",
  "sqlalchemy>=2.0.36",
  "uvicorn[standard]>=0.32.0",
]

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["aurelia_backend"]
